+++ docs/supportscripts/post_request.go
@@ -17,7 +17,7 @@ func main() {
 
 	roletoassume := "arn:aws:iam::248804316466:role/operator"
 	//url := "https://dev.deputy-reporting.api.opg.service.justice.gov.uk/v1/clients/25944618/reports/2d79c411-a8e0-4bda-b9b2-483b1487bc35/supportingdocuments"
-	url := "https://dev.deputy-reporting.api.opg.service.justice.gov.uk/v2/clients/86622299/reports"
+	url := "https://dev.deputy-reporting.api.opg.service.justice.gov.uk/v2/clients/99922299/reports"
 	mysession := session.Must(session.NewSession())
 	creds := stscreds.NewCredentials(mysession, roletoassume)
 	cfg := aws.Config{Credentials: creds,Region: aws.String("eu-west-1")}
