+++ lambda_functions/v1/directory_sha
@@ -1,1 +1,1 @@
-d66f74501623330
+78d06bcfa9c898f
