+++ lambda_functions/v2/directory_sha
@@ -1,1 +1,1 @@
-7e2e00287644ca5
+e700aa2fea2ae36
